+++ config/help.php
@@ -1,0 +1,26 @@
+<?php
+
+return [
+
+    'topics' => [
+        'getting-started' => [
+            'sort' => 1,
+        ],
+        'sharing-and-recipients' => [
+            'sort' => 2,
+        ],
+        'for-recipients' => [
+            'sort' => 3,
+        ],
+        'downloading-files' => [
+            'sort' => 4,
+        ],
+        'approval' => [
+            'sort' => 5,
+        ],
+        'faq' => [
+            'sort' => 6,
+        ],
+    ],
+
+];