+++ resources/views/components/help-link.blade.php
@@ -1,0 +1,10 @@
+@props([
+    'topic',
+])
+
+<a
+    href="{{ route('help.show', ['topic' => $topic]) }}"
+    {{ $attributes->merge(['class' => 'fi-btn-link']) }}
+>
+    @lang('help.learn-more')
+</a>